--- a/app/assets/stylesheets/pages/_home.scss
+++ b/app/assets/stylesheets/pages/_home.scss
@@ -9,7 +9,7 @@ body {
   justify-content: space-evenly;
   align-items: center;
   height: 80vh;
-  touch-action: none;
+  // touch-action: none;
 }
 
 .pill-btn {