--- a/app/assets/stylesheets/pages/_intro.scss
+++ b/app/assets/stylesheets/pages/_intro.scss
@@ -1,7 +1,7 @@
-body {
-  // overflow: hidden;
-  // touch-action: none;
-}
+// body {
+//   // overflow: hidden;
+//   // touch-action: none;
+// }
 
 .intro-body {
   overflow: hidden;
@@ -55,9 +55,9 @@ body {
 
 body { font-family: sans-serif; }
 
-.gallery {
-  // background: #EEE;
-}
+// .gallery {
+//   // background: #EEE;
+// }
 
 .gallery-cell {
   width: 100%;
@@ -72,9 +72,9 @@ body { font-family: sans-serif; }
   max-width: 100%;
 }
 
-.gallery-cell.is-selected {
-  // background: #ED2;
-}
+// .gallery-cell.is-selected {
+//   // background: #ED2;
+// }
 
 .flickity-button {
   display: none !important;