--- a/app/views/products/new.html.erb
+++ b/app/views/products/new.html.erb
@@ -67,7 +67,7 @@
     document.addEventListener('DOMContentLoaded', function() {
       setTimeout(function() {
         document.getElementById('animated-splash-screen').style.display = 'none';
-      }, 3000); // Durée de l'animation en millisecondes
+      }, 2300); // Durée de l'animation en millisecondes
     });
   </script>