--- a/app/views/recycling_spots/index.html.erb
+++ b/app/views/recycling_spots/index.html.erb
@@ -12,6 +12,6 @@
   document.addEventListener('DOMContentLoaded', function() {
     setTimeout(function() {
       document.getElementById('animated-splash-screen').style.display = 'none';
-    }, 1500); // Durée de l'animation en millisecondes
+    }, 1000); // Durée de l'animation en millisecondes
   });
 </script>