--- a/app/views/users/intro.html.erb
+++ b/app/views/users/intro.html.erb
@@ -4,9 +4,9 @@
   <meta name="viewport" content="width=device-width, initial-scale=1.0, maximum-scale=1.0, user-scalable=no"/>
 </head>
 
-<div id="animated-splash-screen">
+<%# <div id="animated-splash-screen">
   <img id="animated-splash-screen" src="/splashscreens/apple-splash-1170x2532.gif" alt="Loading...">
-</div>
+</div> %>
 
 <div class="intro-body particle-container">
 
@@ -39,10 +39,10 @@
   <% end %>
 </div>
 
-<script>
+<%# <script>
   document.addEventListener('DOMContentLoaded', function() {
     setTimeout(function() {
       document.getElementById('animated-splash-screen').style.display = 'none';
     }, 1000); // Durée de l'animation en millisecondes
   });
-</script>
+</script> %>