### Optimized Load Times
`3000ms to 2300ms loading time for camera on scan`
`1500ms to 1000ms loading time for unsorted map while stored in cache`